+++ packages/recoverage/__tests__/sample-package-01/.gitignore
@@ -1,2 +1,3 @@
 node_modules
 coverage
+coverage.sqlite
