+++ packages/recoverage/src/recoverage.ts
@@ -1,0 +1,156 @@
+import type { Hash } from "node:crypto"
+import { createHash } from "node:crypto"
+
+import { $, file, write } from "bun"
+import { Database } from "bun:sqlite"
+import colors from "colors"
+import * as Diff from "diff"
+import { createCoverageMap } from "istanbul-lib-coverage"
+import simpleGit from "simple-git"
+import tmp from "tmp"
+
+const COLUMNS = String(150)
+const DEFAULT_BRANCH = `main`
+
+class BranchCoverage {
+	public git_ref: string
+	public coverage: string
+}
+
+export async function recoverage(): Promise<void> {
+	console.log(`recoverage ------------------------------------`)
+	console.log(`running from`, import.meta.dir)
+	const git = simpleGit(import.meta.dir)
+	const { current: currentGitBranch, branches } = await git.branch()
+
+	const gitStatus = await git.status()
+
+	const mainGitRef = branches[DEFAULT_BRANCH].commit
+	let currentGitRef = branches[currentGitBranch].commit
+	if (!gitStatus.isClean()) {
+		console.log(`git status is not clean`)
+		const gitDiff = await git.diff()
+		const gitStatusHash = createHash(`sha256`).update(gitDiff)
+		const untrackedFileData = await Promise.all(
+			gitStatus.files
+				.filter((f) => f.index === `?`)
+				.map(async (f) => `UNTRACKED: ${await file(f.path).text()}`),
+		)
+		for (const fileData of untrackedFileData) {
+			gitStatusHash.update(fileData)
+		}
+		currentGitRef = `${currentGitRef}-${gitStatusHash.digest(`hex`)}`
+	}
+
+	console.log(`currentGitRef`, currentGitRef)
+	console.log(`mainGitRef`, mainGitRef)
+
+	const coverageFile = file(`./coverage/coverage-final.json`)
+	const coverageJson = await coverageFile.json()
+	const coverageMap = createCoverageMap(coverageJson)
+
+	const db = new Database(`./coverage.sqlite`)
+	function setupDatabase() {
+		return db.run(
+			`create table if not exists coverage (git_ref text, coverage text);`,
+		)
+	}
+	setupDatabase()
+	const insertCoverage = db.prepare(
+		`insert into coverage (git_ref, coverage) values ($git_ref, $coverage)`,
+	)
+	const getCoverage = db
+		.query(`SELECT * FROM coverage WHERE git_ref = $git_ref`)
+		.as(BranchCoverage)
+
+	insertCoverage.run({
+		$git_ref: currentGitRef,
+		$coverage: JSON.stringify(coverageMap),
+	})
+	const mainCoverage = getCoverage.all(mainGitRef)
+	const currentCoverage = getCoverage.all(currentGitRef)
+	console.log({ mainCoverage, currentCoverage })
+
+	const fileList = await $`ls`.text()
+
+	console.log({ fileList })
+
+	if (mainGitRef === currentGitRef) {
+		console.log(`updated main coverage`)
+		process.exit(0)
+	}
+
+	type CoverageEval = {
+		total: number
+		covered: number
+		skipped: number
+		pct: number
+	}
+
+	type JsonSummary = {
+		branches: CoverageEval
+		functions: CoverageEval
+		lines: CoverageEval
+		statements: CoverageEval
+	}
+	type JsonSummaryReport = {
+		[key: string]: JsonSummary | undefined
+	} & {
+		total: JsonSummary
+	}
+
+	async function getCoverageSummary(branchCoverage: BranchCoverage) {
+		const { git_ref, coverage } = branchCoverage
+		const tempDir = tmp.dirSync({ unsafeCleanup: true })
+		await write(`${tempDir.name}/out.json`, coverage)
+
+		const textReport =
+			await $`nyc report --reporter=text --color=0 --temp-dir=${tempDir.name}`
+				.env({ ...process.env, COLUMNS })
+				.text()
+		await $`nyc report --reporter=json-summary --temp-dir=${tempDir.name} --report-dir=${tempDir.name}/coverage`.text()
+		const jsonReport = (await file(
+			`${tempDir.name}/coverage/coverage-summary.json`,
+		).json()) as JsonSummaryReport
+		console.log(`coverage for ${git_ref}:`)
+		console.log(textReport)
+		return { git_ref, textReport, jsonReport }
+	}
+
+	const [mainCoverageSummary, currentCoverageSummary] = await Promise.all([
+		getCoverageSummary(mainCoverage[0]),
+		getCoverageSummary(currentCoverage[0]),
+	])
+
+	const coverageDifference =
+		currentCoverageSummary.jsonReport.total.statements.pct -
+		mainCoverageSummary.jsonReport.total.statements.pct
+
+	function logDiff() {
+		console.log(`coverage diff between ${mainGitRef} and ${currentGitRef}:`)
+		const diff = Diff.diffLines(
+			mainCoverageSummary.textReport,
+			currentCoverageSummary.textReport,
+		)
+		for (const part of diff) {
+			const text = part.added
+				? colors.bgGreen(part.value)
+				: part.removed
+					? colors.bgRed(part.value)
+					: part.value
+			process.stdout.write(text)
+		}
+	}
+
+	if (coverageDifference < 0) {
+		logDiff()
+		console.log(`coverage decreased by ${+coverageDifference}`)
+		process.exit(1)
+	}
+
+	if (coverageDifference > 0) {
+		logDiff()
+		console.log(`coverage increased by ${+coverageDifference}`)
+		process.exit(0)
+	}
+}
